requirements.txt
azureml-core
azureml-sdk
pandas
azureml-widgets
numpy
joblib
scikit-learn
azureml-defaults